# Enhance LinkedList implementation with optimizations