--- a/Golang/algoritms/linked_list.go
+++ b/Golang/algoritms/linked_list.go
@@ -8,46 +8,57 @@ type Node[T comparable] struct {
 }
 
 type LinkedList[T comparable] struct {
-	head *Node[T]
+	head *Node[T] // permite acessar o início da lista em O(1)
+	tail *Node[T] // permite acessar o final da lista em O(1)
+	size int      // rastreia o número de elementos para acesso O(1) ao tamanho
 }
 
 func (l *LinkedList[T]) Unshift(value T) {
 	node := &Node[T]{value, l.head}
 	l.head = node
+	if l.tail == nil {
+		l.tail = node
+	}
+	l.size++
 }
 
 func (l *LinkedList[T]) Append(value T) {
 	node := &Node[T]{value, nil}
 	if l.head == nil {
 		l.head = node
-		return
+		l.tail = node
+	} else {
+		l.tail.next = node
+		l.tail = node
 	}
-	head := l.head
-	for head.next != nil {
-		head = head.next
-	}
-	head.next = node
+	l.size++
 }
 
 func (l *LinkedList[T]) AppendOnIndex(value T, index int) {
+	if index < 0 || index > l.size {
+		fmt.Println("Posição inválida")
+		return
+	}
+
 	if index == 0 {
 		l.Unshift(value)
 		return
 	}
 
-	node := &Node[T]{value, nil}
-	actual := l.head
-	for i := 0; actual != nil && i < index-1; i++ {
-		actual = actual.next
+	if index == l.size {
+		l.Append(value)
+		return
 	}
 
-	if actual == nil {
-		fmt.Println("Posição inválida")
-		return
+	node := &Node[T]{value, nil}
+	actual := l.head
+	for i := 0; i < index-1; i++ {
+		actual = actual.next
 	}
 
 	node.next = actual.next
 	actual.next = node
+	l.size++
 }
 
 func (l *LinkedList[T]) Search(value T) (int, bool) {
@@ -71,9 +82,16 @@ func (l *LinkedList[T]) Delete(value T) bool {
 		if current.value == value {
 			if anterior == nil {
 				l.head = current.next
+				if l.head == nil {
+					l.tail = nil // Lista ficou vazia
+				}
 			} else {
 				anterior.next = current.next
+				if current == l.tail {
+					l.tail = anterior // Atualizando tail se removemos o último elemento
+				}
 			}
+			l.size--
 			return true
 		}
 		anterior = current
@@ -89,4 +107,37 @@ func (l *LinkedList[T]) Debug() {
 		head = head.next
 	}
 	fmt.Println("nil")
+	fmt.Printf("Size: %d\n", l.size)
+}
+
+// O método Size() é uma interface pública para clientes da estrutura
+// ele não deve ser utilizado internamente para um acesso mais eficiente use l.size
+func (l *LinkedList[T]) Size() int {
+	return l.size
+}
+
+func (l *LinkedList[T]) FindByIndex(index int) (T, error) {
+	var zero T
+
+	// Validação de índice usando o campo size
+	if index < 0 || index >= l.size {
+		return zero, fmt.Errorf("índice %d fora dos limites (0-%d)", index, l.size-1)
+	}
+
+	// Otimização: verificar extremidades primeiro
+	if index == 0 {
+		return l.head.value, nil
+	}
+
+	if index == l.size-1 {
+		return l.tail.value, nil
+	}
+
+	// Caso geral: percorrer a lista até o índice desejado
+	current := l.head
+	for i := 0; i < index; i++ {
+		current = current.next
+	}
+
+	return current.value, nil
 }