Título: Otimização da Implementação de Lista Encadeada em Go
## Resumo das Alterações
Este PR implementa otimizações significativas na estrutura de dados LinkedList, melhorando o desempenho e a usabilidade da implementação através da adição de novos recursos e otimização de operações críticas.

## Detalhes Técnicos
#### Otimizações Estruturais
Adicionado ponteiro tail para acesso direto ao último nó em O(1)
Implementado controle de size para rastreamento constante do tamanho da lista
Otimizado algoritmo de inserção no final (Append) utilizando o ponteiro tail

#### Novos Métodos
FindByIndex: Busca por índice com otimização para posições extremas (início/fim)
Size(): Acesso O(1) ao tamanho da lista através de interface pública

#### Melhorias de Performance
Operação Append é 317x mais rápida com o uso do ponteiro tail
Inserção no meio da lista é significativamente mais eficiente para grandes conjuntos
Operações de remoção foram otimizadas para casos especiais

#### Resultados dos Benchmarks
Vantagens da Lista Encadeada vs. Slices
Inserção no início: 2.500x mais rápida (34ns vs 85.263ns)
Menor consumo de memória em operações de inserção no meio
Desempenho estável em inserções, independente do tamanho da estrutura

#### Pontos para Considerar
Slices são mais eficientes em acessos diretos por índice (0.3ns vs 443ns)
Operações de busca por valor são mais rápidas em slices devido à localidade de memória

Para mais informações consulte o arquivo: `Golang/algoritms/result.txt`

#### Testes
Adicionados testes unitários para todos os métodos
Implementados benchmarks comparativos entre LinkedList e slices nativos
Implementada análise detalhada de consumo de recursos (CPU e memória)

#### Conclusões
Esta implementação otimizada de lista encadeada oferece uma alternativa eficiente para casos de uso específicos, como inserções frequentes no início ou no meio da estrutura, com um compromisso de desempenho bem documentado em comparação com slices.